A few minor tweaks:
* Updating the ingredients database with Carafa 1-3
* Changing the populate script so that it compares the precision of floats
  (so that we're not constantly reimporting the same ingredient/style
  attributes because of floating point math).
* Updated the `fermententable` table to store lovibond as a float (not an int).

+++ draughtcraft/model/ingredients.py
@@ -51,7 +51,7 @@ class Fermentable(Ingredient):
     
     type                = Field(Enum(*TYPES))
     ppg                 = Field(Integer)
-    lovibond            = Field(Integer)
+    lovibond            = Field(Float)
     origin              = Field(Enum(*ORIGINS))
 
     additions           = OneToMany('RecipeAddition', inverse='fermentable')
